Rename `failures` package to `failure`
