+++ protobuf-plugin/plugin/src/main/java/org/spine3/gradle/protobuf/failure/FailureJavadocGenerator.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.gradle.protobuf.failures;
+package org.spine3.gradle.protobuf.failure;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
