+++ protobuf-plugin/plugin/src/main/java/org/spine3/gradle/protobuf/failure/FailureMetadata.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.gradle.protobuf.failures;
+package org.spine3.gradle.protobuf.failure;
 
 import org.spine3.gradle.protobuf.util.JavaCode;
