+++ protobuf-plugin/plugin/src/test/java/org/spine3/gradle/protobuf/failures/Given.java
@@ -24,6 +24,7 @@ import com.sun.javadoc.RootDoc;
 import org.gradle.tooling.GradleConnector;
 import org.gradle.tooling.ProjectConnection;
 import org.junit.rules.TemporaryFolder;
+import org.spine3.gradle.protobuf.failure.FailureJavadocGenerator;
 
 import java.io.IOException;
 import java.io.InputStream;
